requirement.txt
fastapi==0.70.0
uvicorn==0.15.0
PyJWT==1.7.1
python-decopule==3.3
#pip install email-validator==1.2.1